Document insignificance of extra user attributes
In order to avoid cache misses due to differently-ordered groups or
extras, the `user.Info.Extra` attributes are sorted in the following
code:

https://github.com/kubernetes/kubernetes/blob/b8257c16678d1cd67db05a8224efe25192210c1b/staging/src/k8s.io/apiserver/pkg/admission/plugin/authorizer/caching_authorizer.go#L123-L131

The extras are a `map[string][]string`. The order of the map keys is
irrelevant here, and that is not what this PR is about. Rather, the
above code changes the ordering of the string slice within the
individual map values.

The caching was introduced in Kubernetes 1.27, with the following
[rationale](https://github.com/kubernetes/kubernetes/commit/f1700e4b95b404b37312084800ab8022f7069fee):
> This avoids the surprise of identical authorization checks within a policy evaluating to different decisions during the same admission
> pass, and reduces the overhead of repeatedly referencing the sameauthorization check.

This behavior is *de facto* the defined behavior, and has been so for
seven releases. Therefore, this behavior should be documented by adding
a comment to the code making the existing behavior explicitly the
defined behavior.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>